[jest] Use FORCE_COLOR to enable integration test colors
The Jest plugin integration tests require colors to be enabled,
which they aren't by default in CI. Previously, the integration
tests changed the `chalk` state to enable colors, but this was
fragile and broke when Yarn hoisted the `chalk` dependency in a
way that caused the integration test to have its own instance of
chalk. Updating that instance's state then had no effect on the
Jest plugin's instance, and the tests failed.

Instead, we can set the FORCE_COLOR environment variable, which
affects all instances of `chalk`. Unfortunately, we have to set
this in package.json rather that in the test because the environment
variable has already been read by the time the test code is loaded.