[cypress] Disable debug colors within Cypress integration tests
The `debug` library only prints timestamped output when terminal colors
are disabled. We disable colors here so that we can better understand the
timing of failed test runs. Note that DEBUG_COLORS is still enabled for
the Jest tests that invoke the Cypress plugin integration tests, since
GitHub Actions lets us show timestamps for each line of that output even
if colors are enabled. However, output from individual test cases is
buffered and printed all at once in the event of a test failure, so the
buffered output needs to include timestamps if we want to understand
the timing within a single Cypress invocation.